+++ libraries/extractor/src/main/java/androidx/media3/extractor/ts/H265Reader.java
@@ -17,6 +17,7 @@ package androidx.media3.extractor.ts;
 
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
+import androidx.media3.common.ColorInfo;
 import androidx.media3.common.Format;
 import androidx.media3.common.MimeTypes;
 import androidx.media3.common.util.Assertions;
@@ -264,6 +265,14 @@ public final class H265Reader implements ElementaryStreamReader {
         .setCodecs(codecs)
         .setWidth(spsData.width)
         .setHeight(spsData.height)
+        .setColorInfo(
+            new ColorInfo.Builder()
+                .setColorSpace(spsData.colorSpace)
+                .setColorRange(spsData.colorRange)
+                .setColorTransfer(spsData.colorTransfer)
+                .setLumaBitdepth(spsData.bitDepthLumaMinus8 + 8)
+                .setChromaBitdepth(spsData.bitDepthChromaMinus8 + 8)
+                .build())
         .setPixelWidthHeightRatio(spsData.pixelWidthHeightRatio)
         .setInitializationData(Collections.singletonList(csdData))
         .build();